+++ SustainU/Views/MapView.swift
@@ -14,15 +14,9 @@ struct MapViewWithOfflinePopup: View {
             MapView(locationManager: locationManager,
                    userTrackingMode: $userTrackingMode,
                    collectionPoints: collectionPoints,
-                   onAnnotationTap: onAnnotationTap,
-                   viewModel: viewModel)
-                .onTapGesture {
-                    viewModel.handleMapTap()
-                }
+                   onAnnotationTap: onAnnotationTap)
             
-            if viewModel.showOfflinePopup {
-                OfflineMapPopupView(isPresented: $viewModel.showOfflinePopup)
-            }
+            OfflineMapPopupView(isPresented: $viewModel.showOfflinePopup)
         }
     }
 }
@@ -32,18 +26,17 @@ struct MapView: UIViewRepresentable {
     @Binding var userTrackingMode: MKUserTrackingMode
     var collectionPoints: [CollectionPoint]
     var onAnnotationTap: (CollectionPoint) -> Void
-    @ObservedObject var viewModel: MapViewModel
+    @StateObject private var viewModel: MapViewModel
     
     init(locationManager: LocationManager,
          userTrackingMode: Binding<MKUserTrackingMode>,
          collectionPoints: [CollectionPoint],
-         onAnnotationTap: @escaping (CollectionPoint) -> Void,
-         viewModel: MapViewModel) {
+         onAnnotationTap: @escaping (CollectionPoint) -> Void) {
         self.locationManager = locationManager
         self._userTrackingMode = userTrackingMode
         self.collectionPoints = collectionPoints
         self.onAnnotationTap = onAnnotationTap
-        self.viewModel = viewModel
+        self._viewModel = StateObject(wrappedValue: MapViewModel(locationManager: locationManager))
     }
     
     func makeCoordinator() -> Coordinator {
@@ -57,9 +50,6 @@ struct MapView: UIViewRepresentable {
         mapView.isPitchEnabled = true
         mapView.isRotateEnabled = true
         mapView.mapType = .standard
-        
-        let tapGesture = UITapGestureRecognizer(target: context.coordinator, action: #selector(Coordinator.handleMapTap(_:)))
-        mapView.addGestureRecognizer(tapGesture)
         
         let compass = MKCompassButton(mapView: mapView)
         compass.compassVisibility = .visible
@@ -126,10 +116,6 @@ struct MapView: UIViewRepresentable {
         
         init(_ parent: MapView) {
             self.parent = parent
-        }
-        
-        @objc func handleMapTap(_ gesture: UITapGestureRecognizer) {
-            parent.viewModel.handleMapTap()
         }
         
         func mapView(_ mapView: MKMapView, viewFor annotation: MKAnnotation) -> MKAnnotationView? {
